Put jagged selection, selection in the same file. Add lj_cuts
